requirements.txt
streamlit
scikit-learn
openai
python-dotenv
joblib
pandas
numpy